backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

python-dotenv==1.0.0
pydantic-settings==2.0.3

supabase==2.4.4

passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0

python-jose[cryptography]==3.3.0
PyJWT==2.7.0

email-validator==2.3.0
pydantic[email]==2.12.5

reportlab==4.0.9
pillow==10.2.0

requests==2.31.0
python-multipart==0.0.6

typing-extensions==4.9.0

